requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==2.0.0
joblib==1.4.2
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
